fix: #78 Cannot switch style to light / dark
The problem is that a required file is only loaded *once*.
This means if you require a file that returns data which
was created using some condition, this condition is not
reevaluated. Instead you have to return a function that
produces the data considering the condition.

Here, instead of just returning the color data when
requireing vscode.colors, I return the function
get_colors(). Then in theme.lua we need to call this
function and - again - make sure it is not only
evaluated once but every time set_highlights() is called.